MI5-69 Odvojene AccountView composable funkcije
AccountView je sadržavao composable funkciju za prikaz liste korisnika i za registraciju. Ove 2 komponente odvojene su u svoje zasebne fileove.
