+++ app/src/main/java/foi/air/szokpt/views/app/AccountView.kt
@@ -2,55 +2,24 @@ package foi.air.szokpt.views.app
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.GridItemSpan
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.rounded.Add
-import androidx.compose.material.icons.rounded.Search
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.SegmentedButton
-import androidx.compose.material3.SegmentedButtonDefaults
-import androidx.compose.material3.SingleChoiceSegmentedButtonRow
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
-import foi.air.szokpt.models.AccountListRole
-import foi.air.szokpt.models.ListedAccountInformation
-import foi.air.szokpt.ui.components.TileSegment
-import foi.air.szokpt.ui.components.interactible_components.FillBouncingButton
-import foi.air.szokpt.ui.components.interactible_components.OutlineBouncingButton
-import foi.air.szokpt.ui.components.list_components.AccountListItem
-import foi.air.szokpt.ui.theme.Alternative
-import foi.air.szokpt.ui.theme.AppBorderRadius
-import foi.air.szokpt.ui.theme.BGLevelOne
-import foi.air.szokpt.ui.theme.BGLevelZeroLow
-import foi.air.szokpt.ui.theme.Primary
-import foi.air.szokpt.ui.theme.Secondary
-import foi.air.szokpt.ui.theme.TextGray
-import foi.air.szokpt.ui.theme.TextWhite
-import foi.air.szokpt.ui.theme.TileSizeMode
-import foi.air.szokpt.views.ROUTE_ALL_ACCOUNT_SEARCH
-import foi.air.szokpt.views.ROUTE_REGISTRATION
+import foi.air.szokpt.ui.components.accounts_components.accountViewComponents.AccountList
+import foi.air.szokpt.ui.components.accounts_components.accountViewComponents.RegisterNewAccount
 
 @Composable
-fun AccountView(navController: NavController){
+fun AccountView(navController: NavController) {
     Column(
         modifier = Modifier
             .fillMaxSize()
@@ -71,7 +40,7 @@ fun AccountView(navController: NavController){
                 .padding(8.dp),
             verticalArrangement = Arrangement.spacedBy(4.dp),
             horizontalArrangement = Arrangement.spacedBy(4.dp)
-        ){
+        ) {
             item(span = { GridItemSpan(2) }) {
                 RegisterNewAccount(navController)
             }
@@ -81,118 +50,3 @@ fun AccountView(navController: NavController){
         }
     }
 }
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun RegisterNewAccount(navController: NavController) {
-    TileSegment(
-        tileSizeMode = TileSizeMode.WRAP_CONTENT,
-        innerPadding = 12.dp,
-        outerMargin = 4.dp,
-        minWidth = 250.dp,
-        minHeight = 20.dp,
-        color = BGLevelOne
-    ) {
-        Column(
-            modifier = Modifier.fillMaxWidth(),
-            verticalArrangement = Arrangement.SpaceBetween,
-        ) {
-            Text(
-                text = "New Account",
-                color = TextWhite,
-                fontSize = 22.sp,
-                fontWeight = FontWeight.SemiBold
-            )
-
-            Column(
-                modifier = Modifier
-                    .fillMaxWidth(),
-                verticalArrangement = Arrangement.SpaceBetween,
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                Spacer(modifier = Modifier.height(8.dp))
-
-                var selectedIndex by remember { mutableStateOf(0) }
-                val options = listOf("User", "Admin")
-                SingleChoiceSegmentedButtonRow {
-                    options.forEachIndexed { index, label ->
-                        val isSelected = index == selectedIndex
-                        SegmentedButton(
-                            shape = SegmentedButtonDefaults.itemShape(index = index, count = options.size),
-                            onClick = { selectedIndex = index },
-                            selected = isSelected,
-                            modifier = Modifier,
-                            colors = SegmentedButtonDefaults.colors(
-                                activeContainerColor = Secondary,
-                                activeContentColor = TextWhite,
-                                activeBorderColor = Primary,
-                                inactiveContainerColor = BGLevelZeroLow,
-                                inactiveContentColor = TextGray,
-                                inactiveBorderColor = TextGray,
-                            )
-                        ) {
-                            Text(
-                                label,
-                                color = if (isSelected) Color.White else Color.Gray,
-                                fontWeight = if (isSelected) FontWeight.Bold else FontWeight.Normal
-                            )
-                        }
-                    }
-                }
-                FillBouncingButton(
-                    modifier = Modifier,
-                    inputText = "Register new ${options[selectedIndex]}",
-                    inputIcon = Icons.Rounded.Add,
-                    buttonColor = Primary,
-                    contentColor = TextWhite,
-                    useSpacerAnimation = true,
-                    useIconAnimation = true
-                ) {
-                    navController.navigate(ROUTE_REGISTRATION + "/${options[selectedIndex]}")
-                }
-            }
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun AccountList(navController: NavController) {
-    TileSegment(
-        tileSizeMode = TileSizeMode.WRAP_CONTENT,
-        innerPadding = 10.dp,
-        outerMargin = 4.dp,
-        minWidth = 250.dp,
-        minHeight = 20.dp,
-        color = BGLevelOne
-    ) {
-        Column(
-            modifier = Modifier.fillMaxWidth(),
-            verticalArrangement = Arrangement.SpaceBetween
-        ) {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Text(
-                    text = "All Accounts",
-                    color = TextWhite,
-                    fontSize = 22.sp,
-                    fontWeight = FontWeight.SemiBold
-                )
-                // TODO: Implement a route to just show a newly created singular account list view
-                //       Called AcccountListView. Where i can only view and search all accounts.
-                OutlineBouncingButton(
-                    modifier = Modifier,
-                    inputText = "",
-                    inputIcon = Icons.Rounded.Search,
-                    contentColor = Primary,
-                    borderColor = Secondary,
-                ) {
-                    navController.navigate(ROUTE_ALL_ACCOUNT_SEARCH)
-                }
-            }
-        }
-    }
-}
-
